fix(life-money-sim): 実装レビュー指摘対応(保存の成否判定・資産額バケットの負値・金額カラムのnumeric化)
- saveResultが成功/失敗を真偽値で返すようにし、SaveButtonは成功時のみ「保存しました」を表示する
  (design.mdのシーケンス図どおり、失敗時は何も表示しない)
- admin集計のASSET_BUCKETS最小区分の下限を-Infinityにし、マイナスの資産額も取りこぼさない
- 金額カラム(monthly_salary等)をinteger→numericに変更し、小数の万円入力を保存できるようにする

+++ supabase/migrations/20260720090000_create_life_money_sim_results.sql
@@ -7,16 +7,16 @@ create table life_money_sim_results (
   created_at timestamptz not null default now(),
   has_spouse boolean not null,
   children_count integer not null,
-  monthly_salary integer not null,
-  personal_expense_monthly integer not null,
-  household_expense_total integer,
-  my_household_share integer,
-  starting_asset integer not null,
+  monthly_salary numeric not null,
+  personal_expense_monthly numeric not null,
+  household_expense_total numeric,
+  my_household_share numeric,
+  starting_asset numeric not null,
   investment_mode boolean not null,
   expected_annual_rate numeric,
   event_count integer not null,
-  final_month_asset integer not null,
-  monthly_surplus integer not null,
+  final_month_asset numeric not null,
+  monthly_surplus numeric not null,
   is_test boolean not null default false
 );
